+++ src/Parsers/SortableDateTime.php
@@ -15,15 +15,15 @@ class SortableDateTime implements ParserInterface
     public function parse($date, $time): ?int
     {
         if ($date > 200000000000) {
-            $y = substr($date, 0, 4);
-            $m = substr($date, 4, 2);
-            $d = substr($date, 6, 2);
-            $h = substr($date, 8, 2);
-            $i = substr($date, 10, 2);
+            $y = (int) substr($date, 0, 4);
+            $m = (int) substr($date, 4, 2);
+            $d = (int) substr($date, 6, 2);
+            $h = (int) substr($date, 8, 2);
+            $i = (int) substr($date, 10, 2);
 
             # The seconds parameter is optional so ensure it is there
             if (strlen($date) > 12) {
-                $s = substr($date, 12, 2);
+                $s = (int) substr($date, 12, 2);
             } else {
                 $s = 0;
             }
